+++ bin/src/help/drush_extra.sh
@@ -1,0 +1,3 @@
+markup_h2 "Drush specific help:"
+output="$( drupal_drush_run -h )"
+echo "${output//drush//bin/drush}"