+++ bin/src/hook_info/composer.sh
@@ -1,0 +1,6 @@
+################################################################################
+# Hook information for the bin/composer script.
+################################################################################
+
+# Common hook information.
+hook_info_run_load "composer.sh"