+++ bin/src/hook_info/hook/init_composer.sh
@@ -1,0 +1,4 @@
+markup_h2 "Install or update local composer"
+markup "${GREY}config/${SCRIPT_NAME}/${RESTORE}init_composer_before${GREY}(_${ENVIRONMENT})${RESTORE}.sh
+${GREY}config/${SCRIPT_NAME}/${RESTORE}init_composer_after${GREY}(_${ENVIRONMENT})${RESTORE}.sh
+"