+++ bin/CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## [0.3.0] - 2015-12-26
+### Added
+- #19 : Added bin/drush command and its documentation.
+- #20 : Added bin/init command and its documentation.
+- #23 : Install composer locally by running the bin/init command.
+- #23 : Install drush locally by running the bin/init command.
+- #30 : Add the configuration option to use local or global composer binary.
+- #32 : Add support to use drush.phar.
+- #34 : Add bin/coder command as a wrapper around the phpcs binary using drupal
+        coding standards.
+- #35 : Add option to force installing 3th party packages (composer, drush,
+        coder, ...).
+
+### Fixed
+- Error when there is no hook info available for a specific command.
+- Added missing documentation about extending the skeleton command help.
+- Added missing help information of wrapped commands like bin/drush and
+  bin/composer.
+- #36 : Fixed missing support for quoted parameters.
+
+
 ## [0.2.0] - 2015-12-02 Documentation
 ### Added
 - #11 Add documentation : documented the drupal-skeleton, how everything works
@@ -44,7 +65,7 @@
 
 
 
-
+[0.3.0]: https://github.com/zero2one/drupal-skeleton/compare/0.2.0...0.3.0
 [0.2.0]: https://github.com/zero2one/drupal-skeleton/compare/0.1.1...0.2.0
 [0.1.2]: https://github.com/zero2one/drupal-skeleton/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/zero2one/drupal-skeleton/compare/0.1.0...0.1.1